MAINT: Fix LGTM.com warning
Multiplication result converted to larger type
Multiplication result may overflow 'int' before it is converted to 'unsigned long'.

One of the multiplication operands should have type 'size_t'.

+++ numpy/core/src/multiarray/nditer_impl.h
@@ -289,7 +289,7 @@ struct NpyIter_AxisData_tag {
         1 + \
         /* intp stride[nop+1] AND char* ptr[nop+1] */ \
         2*((nop)+1) \
-        )*NPY_SIZEOF_INTP )
+        )*(size_t)NPY_SIZEOF_INTP)
 
 /*
  * Macro to advance an AXISDATA pointer by a specified count.
